Solved
You are given a word consisting of lowercase Latin letters. You need to rearrange the letters of the word so that a palindrome results. If there are several possible resulting palindromes, output the lexicographically smallest one (the one that would come first in a standard dictionary). If it is impossible to rearrange the letters into a palindrome, output the word “impossible”.